Split BV validator reporting
Move BV validation metrics, console reporting, and JSON artifact writing out of the main validator so the orchestrator drops from high file-bloat to a tracked advisory warning without touching calculation behavior.

Constraint: scripts/** is ignored, so validate-bv-reporting.ts is force-added as an intentional tracked helper.

Rejected: Ledger-only demotion | the old high finding needed an actual helper boundary and validate:bv proof before being reclassified.

Confidence: high

Scope-risk: narrow

Directive: Keep remaining validate-bv size reductions guarded by BV parity; reference/override loading is the next likely split.

Tested: npm.cmd run validate:bv; npm.cmd run typecheck -- --pretty false; node scripts/qc/validate-maintenance-warning-ledger.mjs; npm.cmd run qc:validate; npm.cmd run qc:app-completion -- --json; npm.cmd run maintain:scan:gate; git diff --check; git diff --cached --check

Not-tested: Full repo-wide legacy maintenance burn-down beyond this validator reporting slice.
